Contribution can enrich existing goals
